fix/카테고리 선택
